rename for sms_reactivation js file
