+++ app/javascript/application.js
@@ -34,4 +34,4 @@ require('./src/sidebar')
 require('./src/tooltip')
 require('./src/session_timeout_poller.js')
 require('./src/casa_org')
-require('./src/sms_reactivation_alert')
+require('./src/sms_reactivation_toggle')
